Now has integrated KY-013 calibration
Default calibration values were close for my uno with the sensor bridge powered from the 5vdc uno pin.  Calibration can be uniquely modified for each Analog pin.  Pin AO calibration value is at EEPROM address 14, A1 at 15 and so on.  that is handled as a signed char, so values above 127 to 256 are actually negative calibration offsets.  The default values are set to 200 which equals -56.  If the sensor is powered by a lower voltage the offsets will probably need to be made closer to 0 (farther from 127, so higher from 200 or lower from values less than 127).